Create a new rails app
Hooray, we've created our first rails up. There's not much to see yet,
just a default view, but the database is up and running and we're good
to go. Speaking of, sqlite uses a file to store database contents. Since
we're going to be pushing this code around from repository to
repository, we might want to keep it from being added